# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/$modelName$": {
            "get": {
                "description": "\"查询列表\"",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "model名称",
                        "name": "dataNames",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "#/definitions/$modelName$",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            },
            "put": {
                "description": "\"更新\"",
                "parameters": [
                    {
                        "description": "#/definitions/$modelName$",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "type": "string",
                        "description": "操作名称[监听子属性使用]",
                        "name": "watchAction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            },
            "post": {
                "description": "\"新增数据\"",
                "parameters": [
                    {
                        "description": "#/definitions/$modelName$",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "新增成功"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/$modelName$/batch": {
            "put": {
                "description": "\"批量更新\"",
                "parameters": [
                    {
                        "description": "#/definitions/$modelName$",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    {
                        "type": "string",
                        "description": "操作名称[监听子属性使用]",
                        "name": "watchAction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "批量更新成功"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            },
            "post": {
                "description": "\"批量新增\"",
                "parameters": [
                    {
                        "description": "#/definitions/$modelName$",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    {
                        "type": "boolean",
                        "description": "是否覆盖已存在",
                        "name": "overwrite",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "批量新增结果",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/fileloader.DataBatchResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            },
            "delete": {
                "description": "\"批量删除\"",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "数据名称",
                        "name": "dataNames",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "批量删除成功"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
...
        "/$modelName$/copy": {
            "post": {
                "description": "\"复制\"",
                "parameters": [
                    {
                        "description": "请求数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fileloader.DataMutation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "复制成功"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/$modelName$/deleteParent/{dataName}": {
            "delete": {
                "description": "\"删除[ByParentDataName]\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "数据名称",
                        "name": "dataName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/$modelName$/export": {
            "get": {
                "description": "\"导出\"",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "dataNames",
                        "name": "dataNames",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "导出成功"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/$modelName$/import": {
            "post": {
                "description": "\"导入\"",
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "导入成功"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
...
        "/$modelName$/rename": {
            "post": {
                "description": "\"重命名\"",
                "parameters": [
                    {
                        "description": "请求数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fileloader.DataMutation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "重命名成功"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/$modelName$/renameParent": {
            "post": {
                "description": "\"重命名[ByParentDataName]\"",
                "parameters": [
                    {
                        "description": "请求数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fileloader.DataMutation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "重命名成功"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/$modelName$/single": {
            "get": {
                "description": "\"查询singleData\"",
                "responses": {
                    "200": {
                        "description": "#/definitions/$modelName$",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/$modelName$/tree": {
            "get": {
                "description": "\"查询树状图\"",
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/fileloader.DataTree"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/$modelName$/withLockUser": {
            "get": {
                "description": "\"带锁查询singleData\"",
                "responses": {
                    "200": {
                        "description": "#/definitions/$modelName$",
                        "schema": {
                            "$ref": "#/definitions/fileloader.DataWithLockUser_data"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/$modelName$/withLockUser/{dataName}": {
            "get": {
                "description": "\"带锁查询\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model名称",
                        "name": "dataName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "#/definitions/$modelName$",
                        "schema": {
                            "$ref": "#/definitions/fileloader.DataWithLockUser_data"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/$modelName$/{dataName}": {
            "get": {
                "description": "\"查询\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model名称",
                        "name": "dataName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "#/definitions/$modelName$",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            },
            "delete": {
                "description": "\"删除\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "数据名称",
                        "name": "dataName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fileloader.DataBatchResult": {
            "type": "object",
            "properties": {
                "dataName": {
                    "type": "string"
                },
                "succeed": {
                    "type": "boolean"
                }
            }
        },
        "fileloader.DataMutation": {
            "type": "object",
            "properties": {
                "dst": {
                    "type": "string"
                },
                "overload": {
                    "type": "boolean"
                },
                "src": {
                    "type": "string"
                }
            }
        },
        "fileloader.DataTree": {
            "type": "object",
            "properties": {
                "extension": {
                    "type": "string"
                },
                "extra": {},
                "isDir": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fileloader.DataTree"
                    }
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "fileloader.DataWithLockUser_data": {
            "type": "object",
            "properties": {
                "data": {},
                "user": {
                    "type": "string"
                }
            }
        },
        "i18n.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/i18n.Errcode"
                        }
                    ]
                },
                "message": {
                    "description": "错误消息",
                    "type": "string"
                },
                "mode": {
                    "description": "业务模块",
                    "type": "string"
                }
            }
        },
...
    }
}